core: fix `runtime-import-check` errors
The code resolves `runtime-import-check` errors present in
`@theia/core`. The changes attempt to not introduce too many breaking
changes by exporting `common` functionalities from the existing files.
