+++ packages/plugin-ext/src/common/plugin-api-rpc.ts
@@ -91,8 +91,7 @@ import type {
     TimelineProviderDescriptor
 } from '@theia/timeline/lib/common/timeline-model';
 import { SerializableEnvironmentVariableCollection } from '@theia/terminal/lib/common/base-terminal-protocol';
-// eslint-disable-next-line @theia/runtime-import-check
-import { ThemeType } from '@theia/core/lib/browser/theming';
+import { ThemeType } from '@theia/core/lib/common/theme';
 import { Disposable } from '@theia/core/lib/common/disposable';
 import { PickOptions, QuickInputButtonHandle, QuickPickItem } from '@theia/core/lib/browser';
